refactor: organize admin module internals and consolidate source roots
Split monolithic exception handlers, response DTOs, and event listeners
into single-responsibility files. Extract extension functions and enums
into dedicated packages. Move shared crypto package-info.java from
java source root to kotlin source root, eliminating the redundant
src/main/java directory.
