+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/advices/BusinessRuleNotFoundExceptionHandler.kt
@@ -1,0 +1,19 @@
+package com.nickdferrara.fitify.admin.internal.advices
+
+import com.nickdferrara.fitify.admin.internal.controller.AdminBusinessRuleController
+import com.nickdferrara.fitify.admin.internal.dtos.response.ErrorResponse
+import com.nickdferrara.fitify.admin.internal.exceptions.BusinessRuleNotFoundException
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice(assignableTypes = [AdminBusinessRuleController::class])
+internal class BusinessRuleNotFoundExceptionHandler {
+
+    @ExceptionHandler(BusinessRuleNotFoundException::class)
+    fun handleBusinessRuleNotFound(ex: BusinessRuleNotFoundException): ResponseEntity<ErrorResponse> {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+            .body(ErrorResponse(ex.message ?: "Business rule not found"))
+    }
+}
