+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/advices/ClassValidationExceptionHandler.kt
@@ -1,0 +1,21 @@
+package com.nickdferrara.fitify.admin.internal.advices
+
+import com.nickdferrara.fitify.admin.internal.controller.AdminClassController
+import com.nickdferrara.fitify.admin.internal.dtos.response.ErrorResponse
+import com.nickdferrara.fitify.admin.internal.dtos.response.ValidationErrorResponse
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.MethodArgumentNotValidException
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice(assignableTypes = [AdminClassController::class])
+internal class ClassValidationExceptionHandler {
+
+    @ExceptionHandler(MethodArgumentNotValidException::class)
+    fun handleValidation(ex: MethodArgumentNotValidException): ResponseEntity<ValidationErrorResponse> {
+        val errors = ex.bindingResult.fieldErrors.associate { it.field to (it.defaultMessage ?: "Invalid value") }
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+            .body(ValidationErrorResponse("Validation failed", errors))
+    }
+}
