+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/advices/InvalidRecurringScheduleExceptionHandler.kt
@@ -1,0 +1,19 @@
+package com.nickdferrara.fitify.admin.internal.advices
+
+import com.nickdferrara.fitify.admin.internal.controller.AdminClassController
+import com.nickdferrara.fitify.admin.internal.dtos.response.ErrorResponse
+import com.nickdferrara.fitify.admin.internal.exceptions.InvalidRecurringScheduleException
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice(assignableTypes = [AdminClassController::class])
+internal class InvalidRecurringScheduleExceptionHandler {
+
+    @ExceptionHandler(InvalidRecurringScheduleException::class)
+    fun handleInvalidSchedule(ex: InvalidRecurringScheduleException): ResponseEntity<ErrorResponse> {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+            .body(ErrorResponse(ex.message ?: "Invalid recurring schedule"))
+    }
+}
