+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/dtos/response/BusinessRuleResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.admin.internal.dtos.response
 
-import com.nickdferrara.fitify.admin.internal.entities.BusinessRule
 import java.time.Instant
 import java.util.UUID
 
@@ -12,14 +11,4 @@ internal data class BusinessRuleResponse(
     val description: String?,
     val updatedBy: String,
     val updatedAt: Instant?,
-)
-
-internal fun BusinessRule.toResponse() = BusinessRuleResponse(
-    id = id!!,
-    ruleKey = ruleKey,
-    value = value,
-    locationId = locationId,
-    description = description,
-    updatedBy = updatedBy,
-    updatedAt = updatedAt,
 )
