+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/dtos/response/OverviewResponse.kt
@@ -1,0 +1,11 @@
+package com.nickdferrara.fitify.admin.internal.dtos.response
+
+import java.time.LocalDate
+import java.util.UUID
+
+internal data class OverviewResponse(
+    val from: LocalDate,
+    val to: LocalDate,
+    val locationId: UUID?,
+    val metrics: List<MetricSummary>,
+)
