+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/exceptions/BusinessRuleNotFoundException.kt
@@ -1,0 +1,4 @@
+package com.nickdferrara.fitify.admin.internal.exceptions
+
+internal class BusinessRuleNotFoundException(ruleKey: String) :
+    RuntimeException("Business rule not found: $ruleKey")
