+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/exceptions/ClassNotFoundException.kt
@@ -1,0 +1,6 @@
+package com.nickdferrara.fitify.admin.internal.exceptions
+
+import java.util.UUID
+
+internal class ClassNotFoundException(classId: UUID) :
+    RuntimeException("Class not found: $classId")
