+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/exceptions/CoachNotFoundException.kt
@@ -1,0 +1,6 @@
+package com.nickdferrara.fitify.admin.internal.exceptions
+
+import java.util.UUID
+
+internal class CoachNotFoundException(coachId: UUID) :
+    RuntimeException("Coach not found: $coachId")
