+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/exceptions/CoachScheduleConflictException.kt
@@ -1,0 +1,6 @@
+package com.nickdferrara.fitify.admin.internal.exceptions
+
+import java.util.UUID
+
+internal class CoachScheduleConflictException(coachId: UUID, details: String) :
+    RuntimeException("Coach $coachId has a schedule conflict: $details")
