+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/exceptions/InvalidBusinessRuleValueException.kt
@@ -1,7 +1,4 @@
 package com.nickdferrara.fitify.admin.internal.exceptions
 
-internal class BusinessRuleNotFoundException(ruleKey: String) :
-    RuntimeException("Business rule not found: $ruleKey")
-
 internal class InvalidBusinessRuleValueException(ruleKey: String, reason: String) :
     RuntimeException("Invalid value for business rule '$ruleKey': $reason")
