+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/exceptions/InvalidRecurringScheduleException.kt
@@ -1,0 +1,4 @@
+package com.nickdferrara.fitify.admin.internal.exceptions
+
+internal class InvalidRecurringScheduleException(reason: String) :
+    RuntimeException("Invalid recurring schedule: $reason")
