+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/exceptions/LocationNotFoundException.kt
@@ -1,0 +1,6 @@
+package com.nickdferrara.fitify.admin.internal.exceptions
+
+import java.util.UUID
+
+internal class LocationNotFoundException(locationId: UUID) :
+    RuntimeException("Location not found: $locationId")
