+++ src/main/kotlin/com/nickdferrara/fitify/admin/internal/service/MetricsService.kt
@@ -5,8 +5,8 @@ import com.nickdferrara.fitify.admin.internal.dtos.response.MetricResponse
 import com.nickdferrara.fitify.admin.internal.dtos.response.MetricSummary
 import com.nickdferrara.fitify.admin.internal.dtos.response.OverviewResponse
 import com.nickdferrara.fitify.admin.internal.dtos.response.TrendComparison
-import com.nickdferrara.fitify.admin.internal.entities.Granularity
-import com.nickdferrara.fitify.admin.internal.entities.MetricType
+import com.nickdferrara.fitify.admin.internal.entities.enums.Granularity
+import com.nickdferrara.fitify.admin.internal.entities.enums.MetricType
 import com.nickdferrara.fitify.admin.internal.entities.MetricsSnapshot
 import com.nickdferrara.fitify.admin.internal.exceptions.InvalidMetricsQueryException
 import com.nickdferrara.fitify.admin.internal.repository.MetricsSnapshotRepository
